% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuyuan.exclude.R
\name{wuyuan.exclude}
\alias{wuyuan.exclude}
\title{To exclude}
\usage{
wuyuan.exclude(df, exclude)
}
\arguments{
\item{df}{result by wuyuan.shanghai}

\item{exclude}{exclude, by fuzzy match}
}
\value{
df
}
\description{
To exclude
}
