% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hosp.wuyuan.R
\name{wuyuan.shanghai}
\alias{wuyuan.shanghai}
\title{To Clear data for wuyuan shagnhai}
\usage{
wuyuan.shanghai(filenames, query)
}
\arguments{
\item{filenames}{filenames}

\item{query}{query words, by fuzzy match}
}
\value{
dataframe
}
\description{
To Clear data for wuyuan shagnhai
}
