% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuyuan.webtool.R
\name{wuyuan.webtool}
\alias{wuyuan.webtool}
\title{Web Tool to Clear Lab Data for Wuyuan Shanghai}
\usage{
wuyuan.webtool()
}
\value{
web
}
\description{
Web Tool to Clear Lab Data for Wuyuan Shanghai
}
